Module solana_stake_program::stake_state

Stake state

delegate stakes to vote accounts

keep track of rewards

own mining pools

Structs

Authorized

CalculatedStakePoints 🔒

CalculatedStakeRewards 🔒

Delegation

Lockup

Meta

PointValue

captures a rewards round as lamports to be awarded and the total points over which those lamports are to be distributed

Stake

ValidatedDelegatedInfo 🔒

After calling validate_delegated_amount(), this struct contains calculated values that are used by the caller.

ValidatedSplitInfo 🔒

After calling validate_split_amount(), this struct contains calculated values that are used by the caller.

Enums

InflationPointCalculationEvent

MergeKind 🔒

SkippedReason

StakeAuthorize

StakeState

Functions

authorize

Authorize the given pubkey to manage stake (deactivate, withdraw). This may be called multiple times, but will implicitly withdraw authorization from the previously authorized staker. The default staker is the owner of the stake account’s pubkey.

authorize_with_seed

authorized_from

calculate_split_rent_exempt_reserve 🔒

calculate_stake_points 🔒

calculate_stake_points_and_credits 🔒

for a given stake and vote_state, calculate how many points were earned (credits * stake) and new value for credits_observed were the points paid

calculate_stake_rewards 🔒

for a given stake and vote_state, calculate what distributions and what updates should be made returns a tuple in the case of a payout of:

create_account

create_account_with_activation_epoch

create_lockup_stake_account

create_stake_history_from_delegations

deactivate

deactivate_delinquent 🔒

delegate

delegation_from

do_create_account 🔒

from

initialize

lockup_from

merge

merge_delegation_stake_and_credits_observed 🔒

meta_from

new_stake 🔒

new_stake_history_entry

null_tracer 🔒

redeem_stake_rewards 🔒

redelegate

redelegate_stake 🔒

set_lockup

split

stake_from

stake_weighted_credits_observed 🔒

Calculate the effective credits observed for two stakes when merging

validate_delegated_amount 🔒

Ensure the stake delegation amount is valid. This checks that the account meets the minimum balance requirements of delegated stake. If not, return an error.

validate_split_amount 🔒

Ensure the split amount is valid. This checks the source and destination accounts meet the minimum balance requirements, which is the rent exempt reserve plus the minimum stake delegation, and that the source account has enough lamports for the request split amount. If not, return an error.

withdraw

Type Definitions

RewriteStakeStatus

StakeActivationStatus